Mock IDOL for tests
